Initial development, more to come.
This import includes initial communication code, including encryption,
for communicating with prjindigo.com's API. It is still incomplete,
but I want to commit it to the repository before going much further.

The primary driver code is still under early development as well,
but contains sensitive and required data to communicate with the
server which needs to be supplied in a different manner before
I add it to the repository.